requirements.txt
# Runtime dependencies
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.58.1
elevenlabs>=1.11.1
pyttsx3>=2.98
